create device

+++ src/API/log.cpp
@@ -1,35 +1,11 @@
 #include "log.h"
 
-VKAPI_ATTR VkBool32 VKAPI_CALL DebugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
-                                             VkDebugUtilsMessageTypeFlagsEXT messageType,
-                                             const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData,
-                                             void* pUserData)
-{
-    switch (messageSeverity) {
-    case VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT:
-        spdlog::error("validation layer: {}", pCallbackData->pMessage);
-        break;
-
-    case VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT:
-        spdlog::warn("validation layer: {}", pCallbackData->pMessage);
-        break;
-
-    default:
-        spdlog::info("validation layer: {}", pCallbackData->pMessage);
-        break;
-    }
-
-    return vk::False;
-}
-
-vk::DebugUtilsMessengerEXT& Log::DebugMessenger()
-{
-    static vk::DebugUtilsMessengerEXT debugMessenger;
-    return debugMessenger;
-}
-
 void Log::CreateDebugMessenger()
 {
+    if (!debug) {
+        return;
+    }
+
     vk::DebugUtilsMessengerCreateInfoEXT createInfo;
     createInfo.setMessageSeverity(
         vk::DebugUtilsMessageSeverityFlagBitsEXT::eVerbose |
@@ -59,5 +35,37 @@ void Log::GetDebugInfo(vk::DebugUtilsMessengerCreateInfoEXT& createInfo)
 
 Log::~Log()
 {
-    Instance::Get().destroyDebugUtilsMessengerEXT(Log::DebugMessenger(), nullptr, Instance::Dldi());
+    if (debug) {
+        Instance::Get().destroyDebugUtilsMessengerEXT(Log::DebugMessenger(), nullptr, Instance::Dldi());
+    }
+}
+
+// getter
+vk::DebugUtilsMessengerEXT& Log::DebugMessenger()
+{
+    static vk::DebugUtilsMessengerEXT debugMessenger;
+    return debugMessenger;
+}
+
+// callback function
+VKAPI_ATTR VkBool32 VKAPI_CALL DebugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
+                                             VkDebugUtilsMessageTypeFlagsEXT messageType,
+                                             const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData,
+                                             void* pUserData)
+{
+    switch (messageSeverity) {
+    case VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT:
+        spdlog::error("validation layer: {}", pCallbackData->pMessage);
+        break;
+
+    case VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT:
+        spdlog::warn("validation layer: {}", pCallbackData->pMessage);
+        break;
+
+    default:
+        spdlog::info("validation layer: {}", pCallbackData->pMessage);
+        break;
+    }
+
+    return vk::False;
 }
